+++ auth0_flutter/android/src/test/kotlin/com/auth0/auth0_flutter/CredentialsManagerMethodCallHandlerTest.kt
@@ -515,4 +515,3 @@ class CredentialsManagerMethodCallHandlerTest {
         verify(mockResult).success(userProfileMap)
     }
 }
-