+++ auth0_flutter/example/windows/flutter/generated_plugin_registrant.cc
@@ -6,6 +6,9 @@
 
 #include "generated_plugin_registrant.h"
 
+#include <auth0_flutter/auth0_flutter_plugin_c_api.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
+  Auth0FlutterPluginCApiRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("Auth0FlutterPluginCApi"));
 }